src/api/pyproject.toml
[tool.poetry]
name = "note-taker-api"
version = "0.1.0"
description = "FastAPI backend for Note-Taker"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = ">=0.114"
uvicorn = {extras = ["standard"], version = ">=0.30"}
pydantic = ">=2.8"
psycopg = {version = ">=3.2", extras = ["binary"]}
SQLAlchemy = ">=2.0"
alembic = ">=1.13"
pydantic-settings = ">=2.4"
requests = ">=2.32"
pyyaml = ">=6.0.1"
structlog = "^25.4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.24"
httpx = ">=0.27"
ruff = ">=0.5"
black = ">=24.0"
mypy = ">=1.10"

[tool.poetry.scripts]
api = "app.main:app"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true


